Settings: Expand all advanced settings by default
A lot of the time, the settings I'm actually looking for are under
"advanced", which slows down my navigation to the target setting by
making me tap "Advanced" to expand it.

These hidden advanced sections are also bad for user discoverability
because it means that the user needs to expand all advanced sections in
order to discover all of the settings. It also makes the user hesitate
and scan through the non-advanced options first to determine whether the
option they're looking for is under advanced, since the advanced vs.
non-advanced grouping is often quite unintuitive. Realistically, with
discoverability and ease of access in mind, I can't think of a situation
in which the advanced sections will actually be helpful.

Change-Id: Ie192f2c680b2fb458f6b7ae01ede4b1ea151e489
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>