Ignore custom app space management activities
Allowing apps to override the "Clear storage" function makes users
powerless when it comes to clearing the data of a rogue app, short of
reinstalling the entire app. I've never seen an app that uses this in a
good way, so just remove it to allow clearing storage of any app.

The most prominent example of this is the Google Search (Velvet) app,
which does not allow the user to clear all data and leaves up to 100 MB
of wasted space even after clearing storage in its custom space manager.
It's also impossible to clear its data after disabling the app, which
forces users to deal with wasted space for no reason.

Change-Id: Ib80d4741750bcbb5c6e1dd844aa763ce4e24854c
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>